+++ src/main/java/com/himalayapp/splanning/server/service/impl/NotificationsServiceImpl.java
@@ -2,9 +2,7 @@ package com.himalayapp.splanning.server.service.impl;
 
 import com.himalayapp.splanning.server.Constants;
 import com.himalayapp.splanning.server.Synchronization;
-import com.himalayapp.splanning.server.entity.DoneTasks;
-import com.himalayapp.splanning.server.entity.Notifications;
-import com.himalayapp.splanning.server.repository.DoneTasksRepository;
+import com.himalayapp.splanning.server.entity.Notification;
 import com.himalayapp.splanning.server.repository.NotificationsRepository;
 import com.himalayapp.splanning.server.repository.SynchronizerRepository;
 import com.himalayapp.splanning.server.service.NotificationsService;
@@ -22,23 +20,28 @@ public class NotificationsServiceImpl implements NotificationsService {
     @Autowired
     private SynchronizerRepository sr;
 
-    public List<Notifications> getAll() {
+    public List<Notification> getAll() {
         return repository.findAll();
     }
 
-    public Notifications save(Notifications enitity, long userId) {
-        Notifications newEntity = repository.saveAndFlush(enitity);
+    public Notification save(Notification enitity, long userId) {
+        System.out.println("save notif" + enitity);
+        Notification newEntity = repository.saveAndFlush(enitity);
 
         int tableId = Constants.dbTables.get(Constants.NOTIFICATIONS_TABLE);
         sr.saveAndFlush(Synchronization.getSynchronizer(newEntity.getId(), tableId, userId));
         return newEntity;
     }
 
-    public void remove(long id) {
-        repository.delete(id);
+    public void remove(long id, long userId) {
+        if (repository.exists(id)) {
+            repository.delete(id);
+            int tableId = Constants.dbTables.get(Constants.NOTIFICATIONS_TABLE);
+            sr.saveAndFlush(Synchronization.getSynchronizer(id, tableId, userId));
+        }
     }
 
-    public Notifications getById(long id) {
+    public Notification getById(long id) {
         return repository.findOne(id);
     }
 }
