+++ src/main/java/com/himalayapp/splanning/server/service/impl/RepeatingServiceImpl.java
@@ -2,9 +2,7 @@ package com.himalayapp.splanning.server.service.impl;
 
 import com.himalayapp.splanning.server.Constants;
 import com.himalayapp.splanning.server.Synchronization;
-import com.himalayapp.splanning.server.entity.DoneTasks;
 import com.himalayapp.splanning.server.entity.Repeating;
-import com.himalayapp.splanning.server.repository.DoneTasksRepository;
 import com.himalayapp.splanning.server.repository.RepeatingRepository;
 import com.himalayapp.splanning.server.repository.SynchronizerRepository;
 import com.himalayapp.splanning.server.service.RepeatingService;
@@ -27,6 +25,7 @@ public class RepeatingServiceImpl implements RepeatingService {
     }
 
     public Repeating save(Repeating enitity, long userId) {
+        System.out.println("save rep" + enitity);
         Repeating newEntity = repository.saveAndFlush(enitity);
 
         int tableId = Constants.dbTables.get(Constants.REPEATING_TABLE);
@@ -34,8 +33,12 @@ public class RepeatingServiceImpl implements RepeatingService {
         return newEntity;
     }
 
-    public void remove(long id) {
-        repository.delete(id);
+    public void remove(long id, long userId) {
+        if (repository.exists(id)) {
+            repository.delete(id);
+            int tableId = Constants.dbTables.get(Constants.REPEATING_TABLE);
+            sr.saveAndFlush(Synchronization.getSynchronizer(id, tableId, userId));
+        }
     }
 
     public Repeating getById(long id) {
